Improved motion detector

+++ app/src/main/rs/motion.rs
@@ -3,11 +3,11 @@
 #pragma rs_fp_relaxed
 #pragma rs reduce(rate) accumulator(addDistance) combiner(sum)
 
-static void addDistance(float *accum, uchar4 a, uchar4 b) {
+static void addDistance(double *accum, uchar4 a, uchar4 b) {
     uchar4 d = max(a, b) - min(a, b);
     *accum += d.x + d.y + d.z;
 }
 
-static void sum(float *accum, const float *val) {
+static void sum(double *accum, const double *val) {
     *accum += *val;
 }
